requirements.txt
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.1
pyspark
findspark
kafka-python
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
mysql-connector-python
nltk
streamlit
plotly
flask
dash
dash_bootstrap_components
pymysql